+++ src/Marketplace/interfaces/IGauge.sol
@@ -3,9 +3,9 @@
 pragma solidity 0.8.23;
 
 interface IGauge {
-    function notifyRewardAmount(uint amount) external;
+    function notifyRewardAmount(uint256 amount) external;
     function getReward(address account) external;
-    function claimFees() external returns (uint claimed0, uint claimed1);
-    function left() external view returns (uint);
+    function claimFees() external returns (uint256 claimed0, uint256 claimed1);
+    function left() external view returns (uint256);
     function isForPair() external view returns (bool);
 }
